ignoring mma broken, trying to fix mean issue

+++ src/boot/ksdk1.1.0/boot.c
@@ -1670,7 +1670,7 @@ main(void)
 		mean_x=combined_x/cycles;
 		mean_y=combined_y/cycles;
 		mean_z=combined_z/cycles;
-		warpPrint("Mean x %d,", floor(mean_x));
+		warpPrint("Mean x %d,",(int) floor(mean_x));
 	#endif
 
 	#if (WARP_BUILD_ENABLE_DEVINA219)
